+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 PyBuilder SemVer Git Tag Plugin [![Build Status](https://travis-ci.org/AlexeySanko/pybuilder_semver_git_tag.svg?branch=master)](https://travis-ci.org/AlexeySanko/pybuilder_semver_git_tag)
 =======================
-1.0.8
+1.1.0
 ---
+- `coverage<4.4.2` - wait PyBuilder fix
+- add default initialization on import stage (work with command line properties)
 - `__version__` added
 
 1.0.7